services_scaffold/pose_estimation_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
python-dotenv==1.0.0
boto3==1.33.1
httpx==0.25.2
